requirements.txt
matplotlib==2.1.1
numpy==1.13.3
pandas==1.2.3
pillow==8.1.2
scikit-learn==0.24.1
pytorch==1.8.1
torchvision==0.9.1
scikit-image
